:zap: improve performance
