+++ src/pages/solidText/utils/constants.ts
@@ -3,7 +3,6 @@ export const MATRIX_SIZE = 100;
 
 export const THETA_NUM = 150;
 export const PHI_NUM = 300;
-export const THETA_THRESHOLD = 0.1;
 
 export const majorRadius = 30;
 export const minorRadius = 10;
